# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BeeKeeper API",
        "description": "Documentation OpenAPI des endpoints backend BeeKeeper.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Serveur courant"
        }
    ],
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Inscrire un utilisateur",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur cree avec token Sanctum",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Connecter un utilisateur",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion reussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Recuperer l utilisateur connecte",
                "operationId": "me",
                "responses": {
                    "200": {
                        "description": "Utilisateur authentifie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifie"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "delete": {
                "tags": [
                    "Auth"
                ],
                "summary": "Deconnecter l utilisateur courant",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Token courant revoque",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifie"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ruchers": {
            "get": {
                "tags": [
                    "Ruchers"
                ],
                "summary": "Lister les ruchers de l utilisateur connecte",
                "operationId": "listRuchers",
                "responses": {
                    "200": {
                        "description": "Liste des ruchers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Rucher"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifie"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Ruchers"
                ],
                "summary": "Creer un rucher",
                "operationId": "storeRucher",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreRucherRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rucher cree",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RucherWriteResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifie"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ruchers/{rucher}": {
            "get": {
                "tags": [
                    "Ruchers"
                ],
                "summary": "Afficher un rucher",
                "operationId": "showRucher",
                "parameters": [
                    {
                        "name": "rucher",
                        "in": "path",
                        "description": "ID du rucher",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rucher trouve",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Rucher"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifie"
                    },
                    "403": {
                        "description": "Acces refuse"
                    },
                    "404": {
                        "description": "Rucher introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Ruchers"
                ],
                "summary": "Mettre a jour un rucher",
                "operationId": "updateRucher",
                "parameters": [
                    {
                        "name": "rucher",
                        "in": "path",
                        "description": "ID du rucher",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRucherRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rucher mis a jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RucherWriteResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifie"
                    },
                    "403": {
                        "description": "Acces refuse"
                    },
                    "404": {
                        "description": "Rucher introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Ruchers"
                ],
                "summary": "Supprimer un rucher",
                "operationId": "destroyRucher",
                "parameters": [
                    {
                        "name": "rucher",
                        "in": "path",
                        "description": "ID du rucher",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rucher supprime",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifie"
                    },
                    "403": {
                        "description": "Acces refuse"
                    },
                    "404": {
                        "description": "Rucher introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Alice Martin"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "alice@example.com"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Rucher": {
                "required": [
                    "id",
                    "user_id",
                    "name",
                    "localisation",
                    "nb_emplacements"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Rucher des Tilleuls"
                    },
                    "localisation": {
                        "type": "string",
                        "example": "Lyon"
                    },
                    "description": {
                        "type": "string",
                        "example": "Rucher principal",
                        "nullable": true
                    },
                    "nb_emplacements": {
                        "type": "integer",
                        "example": 12,
                        "minimum": 0
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "required": [
                    "name",
                    "email",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Alice Martin"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "alice@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "password123"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "format": "password",
                        "example": "password123"
                    }
                },
                "type": "object"
            },
            "LoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "alice@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "password123"
                    }
                },
                "type": "object"
            },
            "StoreRucherRequest": {
                "required": [
                    "name",
                    "localisation",
                    "nb_emplacements"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Rucher des Tilleuls"
                    },
                    "localisation": {
                        "type": "string",
                        "example": "Lyon"
                    },
                    "description": {
                        "type": "string",
                        "example": "Rucher principal",
                        "nullable": true
                    },
                    "nb_emplacements": {
                        "type": "integer",
                        "example": 12,
                        "minimum": 0
                    }
                },
                "type": "object"
            },
            "UpdateRucherRequest": {
                "required": [
                    "name",
                    "localisation",
                    "nb_emplacements"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Rucher du Verger"
                    },
                    "localisation": {
                        "type": "string",
                        "example": "Grenoble"
                    },
                    "description": {
                        "type": "string",
                        "example": "Mise a jour du rucher",
                        "nullable": true
                    },
                    "nb_emplacements": {
                        "type": "integer",
                        "example": 16,
                        "minimum": 0
                    }
                },
                "type": "object"
            },
            "AuthResponse": {
                "required": [
                    "message",
                    "user",
                    "token"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Connexion reussie"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "token": {
                        "type": "string",
                        "example": "1|aVeryLongSanctumToken"
                    }
                },
                "type": "object"
            },
            "RucherWriteResponse": {
                "required": [
                    "message",
                    "data"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Rucher cree"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Rucher"
                    }
                },
                "type": "object"
            },
            "MessageResponse": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Operation effectuee"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token Laravel Sanctum a envoyer dans le header Authorization: Bearer {token}.",
                "bearerFormat": "Bearer token",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Endpoints d authentification"
        },
        {
            "name": "Ruchers",
            "description": "Gestion des ruchers utilisateurs"
        }
    ]
}